refactor terms
