--- example-app/SantokuApp/src/features/terms/client-states/useAgreed.ts
@@ -1,6 +1,0 @@
-import {useClientState} from 'bases/react-query/useClientState';
-import {TermsOfServiceAgreementStatus} from 'features/backend/apis/model';
-
-export const useAgreed = () => {
-  return useClientState<TermsOfServiceAgreementStatus>(['terms', 'agreedStatus']);
-};
