--- example-app/SantokuApp/src/features/terms/client-states/useButtonDisable.ts
@@ -1,5 +1,0 @@
-import {useClientState} from 'bases/react-query/useClientState';
-
-export const useButtonDisable = () => {
-  return useClientState(['terms', 'buttonDisable'], true);
-};
