--- example-app/SantokuApp/src/features/terms/client-states/useIsExited.ts
@@ -1,5 +1,0 @@
-import {useClientState} from 'bases/react-query/useClientState';
-
-export const useIsExited = () => {
-  return useClientState(['terms', 'isExited'], false);
-};
