--- example-app/SantokuApp/src/features/terms/client-states/useIsWebViewError.ts
@@ -1,5 +1,0 @@
-import {useClientState} from 'bases/react-query/useClientState';
-
-export const useIsWebViewError = () => {
-  return useClientState(['terms', 'isWebViewError'], false);
-};
