--- example-app/SantokuApp/src/features/terms/use-cases/useOnScrollEndOnce.ts
@@ -1,9 +1,0 @@
-import {useCallback} from 'react';
-
-import {useButtonDisable} from '../client-states/useButtonDisable';
-
-export const useOnScrollEndOnce = () => {
-  const [, setButtonDisable] = useButtonDisable();
-  const onScrollEndOnce = useCallback(() => setButtonDisable(false), [setButtonDisable]);
-  return {onScrollEndOnce};
-};
