--- example-app/SantokuApp/src/features/terms/use-cases/useResetWebViewError.ts
@@ -1,11 +1,0 @@
-import {useCallback} from 'react';
-
-import {useIsWebViewError} from '../client-states/useIsWebViewError';
-
-export const useResetWebViewError = () => {
-  const [, setIsWebViewError] = useIsWebViewError();
-  const resetWebViewError = useCallback(() => {
-    setIsWebViewError(false);
-  }, [setIsWebViewError]);
-  return {resetWebViewError};
-};
